Refactor and fix bugs
1. Add SleepMutex to kernel-sync
2. Init rust-vfs
3. Fix bugs in fatfs: drop called without lock. Now global lock is applied for further refactor (preemption).

+++ crates/signal-defs/src/sigset.rs
@@ -1,5 +1,6 @@
+///
 #[derive(Debug, Default, Clone, Copy)]
-pub struct SigSet(usize);
+pub struct SigSet(u64);
 
 impl SigSet {
     /// Creates a new `SigSet`.
@@ -48,8 +49,8 @@ impl SigSet {
     }
 }
 
-impl From<usize> for SigSet {
-    fn from(value: usize) -> Self {
+impl From<u64> for SigSet {
+    fn from(value: u64) -> Self {
         Self(value)
     }
 }
